requirements.txt
PyYAML
telebot
aiohttp
retry
loguru